// E_commerce-project-main/src/types/forms.ts
export interface LoginFormValues {
  email: string;
  password: string;
}

export interface SignupFormValues {
  name: string;
  email: string;
  password: string;
  rePassword: string;
  phone: string;
}
export interface paymentFormValues {
  city: string;
  streetAddress: string;
  phone: string;
}